Inline logic in ProvTable
