+++ context/app/static/js/components/ProvTable.jsx
@@ -2,13 +2,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import Divider from '@material-ui/core/Divider';
-import ProvTypesList from './ProvTypesList';
+import List from '@material-ui/core/List';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemText from '@material-ui/core/ListItemText';
 
 const FlexContainer = styled.div`
     display: flex;
     flex-grow: 1;
     justify-content:space-around;
 `;
+
+function ListItemLink(props) {
+  // eslint-disable-next-line react/jsx-props-no-spreading
+  return <ListItem button component="a" {...props} />;
+}
 
 function ProvTable(props) {
   const { provData, typesToSplit } = props;
@@ -20,17 +27,37 @@ function ProvTable(props) {
 
   return (
     <FlexContainer>
-      {types.map((type, i) => (
-        type && type.length
-          ? (
-            <React.Fragment key={`provenance-list-${typesToSplit[i].toLowerCase()}`}>
-              <ProvTypesList data={type} />{
-                // eslint-disable-next-line react/no-array-index-key
-                i < (types.length - 1) && <Divider key={`provenance-table-divider-${i}`} orientation="vertical" flexItem />
-                }
-            </React.Fragment>
-          )
-          : (null)))}
+      <List>
+        {types[0].map((item) => (
+          <ListItemLink key={item['hubmap:uuid']} href={`/browse/dataset/${item['hubmap:uuid']}`}>
+            <ListItemText primary={`${item['prov:type']}: ${item['hubmap:displayDOI']}`} />
+          </ListItemLink>
+        ))}
+      </List>
+      <Divider orientation="vertical" flexItem />
+      <List>
+        {
+          types[1]
+            ? types[1].map((item) => (
+              <ListItemLink key={item['hubmap:uuid']} href={`/browse/dataset/${item['hubmap:uuid']}`}>
+                <ListItemText primary={`${item['prov:type']}: ${item['hubmap:displayDOI']}`} />
+              </ListItemLink>
+            ))
+            : 'TODO'
+        }
+      </List>
+      <Divider orientation="vertical" flexItem />
+      <List>
+        {
+          types[2]
+            ? types[2].map((item) => (
+              <ListItemLink key={item['hubmap:uuid']} href={`/browse/dataset/${item['hubmap:uuid']}`}>
+                <ListItemText primary={`${item['prov:type']}: ${item['hubmap:displayDOI']}`} />
+              </ListItemLink>
+            ))
+            : 'TODO'
+        }
+      </List>
     </FlexContainer>
   );
 }
